moved rest of tests into one file
