+++ api/data_refinery_api/test/test_compendia.py
@@ -3,7 +3,7 @@ import json
 from django.urls import reverse
 from rest_framework.test import APITestCase
 
-from data_refinery_api.tests import API_VERSION
+from data_refinery_api.test.test_api_general import API_VERSION
 from data_refinery_common.models import (
     ComputationalResult,
     ComputedFile,
