+++ api/data_refinery_api/test/test_sentry_middleware.py
@@ -5,7 +5,7 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from data_refinery_api.tests import API_VERSION
+from data_refinery_api.test.test_api_general import API_VERSION
 from data_refinery_api.views import DatasetView, ExperimentListView
