+++ api/data_refinery_api/test/test_stats.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 from django.urls import reverse
 from rest_framework.test import APITestCase
 
-from data_refinery_api.tests import API_VERSION
+from data_refinery_api.test.test_api_general import API_VERSION
 
 
 MOCK_NOMAD_RESPONSE = [
